Revert "Rename Defines.php to CommonSettings.php (#4683)"
This reverts commit 2d4bdacc024ad2c8ef99bd55678b433c6d7d9b60.
