+++ LocalSettings.php
@@ -5196,7 +5196,7 @@ if ( wfHostname() === 'test101' ) {
 }
 
 // Define last to avoid all dependencies
-require_once '/srv/mediawiki/config/CommonSettings.php';
+require_once '/srv/mediawiki/config/Defines.php';
 require_once '/srv/mediawiki/config/LocalWiki.php';
 
 // Define last - Extension message files for loading extensions
